add .env for mongo, nest ReservationTable component into admin_dashboard, add endpoint for ReservationTable, define endpoint in server, make api call to getAllAppointments
