+++ client/app/admin-dashboard/page.jsx
@@ -1,9 +1,16 @@
-import AdminDashboard from "../../components/AdminDashboard";
+"use client";
+import { useState, useEffect } from "react";
+import ReservationTable from "../../components/ReservationTable";
 
 export default function AdminDashboardView() {
+  const [appointments, setAppointments] = useState([]);
+
   return (
     <div>
-      <AdminDashboard />
+      <ReservationTable
+        appointments={appointments}
+        setAppointments={setAppointments}
+      />
     </div>
   );
 }
