+++ client/package.json
@@ -12,14 +12,16 @@
     "@emotion/cache": "^11.13.1",
     "@emotion/react": "^11.13.3",
     "@emotion/styled": "^11.13.0",
-    "@mui/icons-material": "^6.1.7",
     "@fontsource/roboto": "^5.1.0",
+    "@mui/icons-material": "^6.1.7",
     "@mui/material": "^6.1.6",
     "@mui/material-nextjs": "^6.1.6",
     "@mui/x-data-grid": "^7.22.2",
     "@mui/x-date-pickers": "^7.22.2",
     "dayjs": "^1.11.13",
     "joi": "^17.13.3",
+    "mongodb": "^6.10.0",
+    "mongoose": "^8.8.1",
     "next": "15.0.3",
     "react": "^18.3.1",
     "react-dom": "^18.3.1"
